SPRINT8: multer como middleware
