--- src/middlewares/multer.js
@@ -1,3 +1,0 @@
-
-const multer = require('multer')
-module.exports = multer;
