+++ src/routes/products.js
@@ -1,29 +1,17 @@
-const path = require('path');
 const express = require('express');
-const multer = require('../middlewares/multer');
+const multerProductsMiddleware = require('../middlewares/multerProductsMiddleware');
 const resizeImagesMiddleware = require('../middlewares/resizeImagesMiddleware');
 const router = express.Router();
 const productController = require('../controllers/productController');
 const userTypeMiddleware = require('../middlewares/userTypeMiddleware');
 const { productsCreateValidations, productsEditValidations } = require('../validations/productsValidations');
 
-const storage = multer.memoryStorage();
-function fileFilter(req, file, cb) {
-    const acceptedFileExtensions = [".jpg", ".png", ".jpeg"];
-    const isAccepted = acceptedFileExtensions.includes(path.extname(file.originalname));
-    if (!isAccepted) {
-        req.file = file;
-    }
-    cb(null, isAccepted);
-}
-const uploadFile = multer({ storage: storage, fileFilter: fileFilter });
-
 router.get('/', productController.productList);
 router.get('/product-detail/:id', productController.productDetail);
 router.get('/create', userTypeMiddleware, productController.create);
-router.post('/create', uploadFile.single('img'), resizeImagesMiddleware, productsCreateValidations, productController.saveNewProduct);
+router.post('/create', multerProductsMiddleware.single('img'), resizeImagesMiddleware, productsCreateValidations, productController.saveNewProduct);
 router.get('/edit/:id', userTypeMiddleware, productController.editar);
-router.put('/edit/:id', uploadFile.single('img'), resizeImagesMiddleware, productsEditValidations, productController.actualizar);
+router.put('/edit/:id', multerProductsMiddleware.single('img'), resizeImagesMiddleware, productsEditValidations, productController.actualizar);
 router.delete('/delete/:id', productController.delete);
 router.post('/:id/add_to_cart', productController.addToCart);
 router.get('/cart', productController.getCart);
